[CELEBORN-174][REFACTOR] Move AppDiskUsage related to meta package
